Fix swauth default endpoint
After moving the account management function to the management node,
the default endpoint was set to the management node.  The UI should be
using the proxy server instead, as it hosts the ui, or the vip on
which the ui is hosted in a multi-proxy scenario.  Swauth, on the
other hand, needs to use the management server as account management
is disabled.

In a multiple proxy cluster, `swift_swauth_public_endpoint` should be
set to the vip as a variable override.  In the default config for easy
provisioning of cloud servers, the default values are now as follows:

swift_swauth_cluster_domain: "{{hostvars[prefix + '-proxy']['ansible_eth0']['ipv4']['address']}}"
swift_swauth_public_endpoint: "http://{{swift_swauth_cluster_domain}}:80/v1"
swift_swauth_default_cluster:  "local#{{swift_swauth_public_endpoint}}#http://127.0.0.1:8080/v1"

The `swauth_cluster_domain` will be set to the public ip of the first
proxy server.  `swift_swauth_public_endpoint` will use that server as
the main endpoint, which will be respected by the ui.  Swauth will use
the "internal" url, specified after the second #, which will always go
to localhost.

This results in the desired behavior: The ui now goes to the CORS
appropriate proxy.  Swauth commands work on the management node.
Swauth commands fail on the proxy node.